Paratize v2.0.0 → Overview

Paratize v2.0.0

Modules

Modules summary

Paratize
Provides a set of functions that does parallel processing on collection of data or functions

Paratize.Base
Provides the base implementation for parallel_each/3 and parallel_map/3

Paratize.Chunk
Implementation of parallel exec via chunks and Task.async / Task.await

Paratize.Pool
Implementation of parallel exec via pool of spawned workers processes

Paratize.TaskOptions
Struct holding the configurations for executing the workload in parallel