Paratize v2.1.4 Paratize.TaskOptions

Struct holding the configurations for executing the workload in parallel.

Types

t()

t() :: %Paratize.TaskOptions{size: :schedulers | non_neg_integer, timeout: :infinity | non_neg_integer}

Functions

__struct__()

size - number of workers, default: the number of CPU cores given by :erlang.system_info(:schedulers).

timeout - timeout in ms, integer, default: 5000, exit(:timeout,…) if no result is return by any of the workers within the period. To disable timeout, use :infinity.

worker_count(task_options)

worker_count(t) :: non_neg_integer

Returns the actual worker count based on %Paratize.TaskOptions{}.