Paratize v2.1.0 → Overview → Paratize → TaskOptions

Paratize.TaskOptions

Summary

Functions

Struct holding the configurations for executing the workload in parallel.

Source

Summary

__struct__()

size - number of workers, default: the number of CPU cores given by :erlang.system_info(:schedulers).

timeout - timeout in ms, integer, default: 5000, exit(:timeout,…) if no result is return by any of the workers within the period. To disable timeout, use :infinity

Functions

__struct__()
(function) # ↑

Specs:

__struct__ :: %Paratize.TaskOptions{size: term, timeout: term}

size - number of workers, default: the number of CPU cores given by :erlang.system_info(:schedulers).

timeout - timeout in ms, integer, default: 5000, exit(:timeout,…) if no result is return by any of the workers within the period. To disable timeout, use :infinity.

Source